+++ src/ModFramework Example Mod/engineering/obj_content_injector.lua
@@ -12,4 +12,13 @@ function create(q,v_modid)
 		--Since we dont want to add new items on loaded saves we return the function here.
 		return
 	end
+
+	--load needed types
+	local baseInjectors = Mod.Types.BaseInjectors
+
+	--If we want to add a base mech we can use the Mod.Types.BaseInjectors references to add the correct one
+	Mod.Engineering.AddInjector(baseInjectors.LowPressure)
+	Mod.Engineering.AddInjector(baseInjectors.LowPressure)
+	Mod.Engineering.AddInjector(baseInjectors.HighPressure)
+	Mod.Engineering.AddInjector(baseInjectors.HighPressure)
 end